% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{languages}
\alias{languages}
\title{Language information}
\format{A tibble with 2 rows and 4 variables.}
\usage{
data(languages)
}
\description{
A dataset containing the names, date orders, simplifications,
and replacements of 2 languages.
}
\keyword{datasets}
