% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midas}
\alias{midas}
\title{MIDAS Standardization Examples}
\format{A vector of length 8115.}
\usage{
data(midas)
}
\description{
A dataset containing eight thousand standardization examples
of the MIDAS (Marburger Informations-, Dokumentations- und
Administrations-System).
}
\keyword{datasets}
