% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schemes}
\alias{schemes}
\title{Language-Specific Scheme Variants}
\format{A tibble with 3583 rows and 3 variables.}
\usage{
data(schemes)
}
\description{
A dataset containing the values, schemes, and languages for
over three thousand language-specific scheme variants.
}
\keyword{datasets}
