% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstruwwel.R
\name{unstruwwel}
\alias{unstruwwel}
\title{Detect and Parse Historic Dates}
\usage{
unstruwwel(x, language = "en", midas = FALSE, ...)
}
\arguments{
\item{x}{Input vector. Either a character vector, or something
coercible to one.}

\item{language}{Language code of the input vector as defined in
ISO 639-1. Defaults to `en`, i.e, English.}

\item{midas}{If `TRUE`, input vector was standardized using MIDAS
(Marburger Informations-, Dokumentations- und Administrations-
System). See \url{https://doi.org/10.11588/artdok.00003770}.}

\item{...}{Optional arguments (currently not used).}
}
\value{
A named list of vectors.
}
\description{
Detect and Parse Historic Dates, e.g., to ISO 8601:2-2019.
}
\examples{
unstruwwel("1. Hälfte 19. Jahrhundert", language = "de")

}
