% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-century.R
\docType{class}
\name{Century}
\alias{Century}
\title{Set a Century and Get its Time Interval}
\description{
Set a Century and Get its Time Interval

Set a Century and Get its Time Interval
}
\details{
An Object of \code{\link{R6Class}} with methods to set
common time periods and specifications for centuries.
}
\examples{
\donttest{
  x <- Century$new(15)
  x$take(2, type = "third")
}

}
\keyword{date}
\keyword{time}
\section{Super class}{
\code{\link[unstruwwel:Period]{unstruwwel::Period}} -> \code{Century}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Century$new()}}
\item \href{#method-clone}{\code{Century$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="unstruwwel" data-topic="Period" data-id="set_fuzzy">}\href{../../unstruwwel/html/Period.html#method-set_fuzzy}{\code{unstruwwel::Period$set_fuzzy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="unstruwwel" data-topic="Period" data-id="take">}\href{../../unstruwwel/html/Period.html#method-take}{\code{unstruwwel::Period$take()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Helper function to specify the beginning of a century.


Helper function to specify the middle of a century.


Helper function to specify the end of a century.


Create a century.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Century$new(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numerical scalar.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object of \code{\link{R6Class}} with methods to set
common time periods and specifications for centuries.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Century$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
