% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-century.R
\docType{class}
\name{Century}
\alias{Century}
\title{Set a Century and Get its Time Interval}
\format{An \code{\link{R6Class}} object.}
\usage{
Century
}
\value{
Object of \code{\link{R6Class}} with methods to set
common time periods and specifications for centuries.
}
\description{
Set a Century and Get its Time Interval
}
\section{Fields}{

\describe{
\item{\code{interval}}{Stores a time interval.}
}}

\section{Arguments}{

\describe{
  \item{\code{x}}{A numerical scalar. The range of valid values
    depends on \code{type}. If \code{type} is \code{"early"},
    \code{"mid"}, or \code{"late"}, \code{x} is ignored.}
  \item{\code{type}}{A character scalar. The following values
    are supported: \code{"early"}, \code{"mid"}, \code{"late"},
    \code{"quarter"}, \code{"third"}, and \code{"half"}. If
    \code{type} is `NULL`, \code{x} defines a decade.}
}
}

\section{Methods}{

\describe{
  \item{\code{take()}}{Set a time period or specification.}
}
}

\examples{
\donttest{
  x <- Century$new(15)
  x$take(2, type = "third")
}

}
\keyword{date}
\keyword{time}
