% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-decade.R
\docType{class}
\name{Decade}
\alias{Decade}
\title{Set a Decade and Get its Time Interval}
\description{
Set a Decade and Get its Time Interval

Set a Decade and Get its Time Interval
}
\details{
An Object of \code{R6Class} with methods to set
common time periods and specifications for decades.
}
\examples{
if (interactive()) {
x <- Decade$new(1520)
x$take(1, type = "half")
}

}
\section{Super class}{
\code{\link[unstruwwel:Periods]{unstruwwel::Periods}} -> \code{Decade}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Decade-new}{\code{Decade$new()}}
\item \href{#method-Decade-clone}{\code{Decade$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="unstruwwel" data-topic="Periods" data-id="set_additions"><a href='../../unstruwwel/html/Periods.html#method-Periods-set_additions'><code>unstruwwel::Periods$set_additions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="unstruwwel" data-topic="Periods" data-id="take"><a href='../../unstruwwel/html/Periods.html#method-Periods-take'><code>unstruwwel::Periods$take()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decade-new"></a>}}
\if{latex}{\out{\hypertarget{method-Decade-new}{}}}
\subsection{Method \code{new()}}{
Helper function to specify the beginning of a decade.


Helper function to specify the middle of a decade.


Helper function to specify the end of a decade.


Create a decade.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decade$new(value, official_def = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numerical scalar.}

\item{\code{official_def}}{If `TRUE`, the official definition that
begins with the year 1 is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object of \code{R6Class} with methods to set
common time periods and specifications for decades.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decade-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Decade-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decade$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
