% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-decade.R
\docType{class}
\name{Decade}
\alias{Decade}
\title{Set a Decade and Get its Time Interval}
\description{
Set a Decade and Get its Time Interval

Set a Decade and Get its Time Interval
}
\details{
An Object of \code{\link{R6Class}} with methods to set
common time periods and specifications for decades.
}
\examples{
\donttest{
  x <- Decade$new(1520)
  x$take(1, type = "half")
}

}
\keyword{date}
\keyword{time}
\section{Super class}{
\code{\link[unstruwwel:Period]{unstruwwel::Period}} -> \code{Decade}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Decade$new()}}
\item \href{#method-clone}{\code{Decade$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="unstruwwel" data-topic="Period" data-id="set_additions">}\href{../../unstruwwel/html/Period.html#method-set_additions}{\code{unstruwwel::Period$set_additions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="unstruwwel" data-topic="Period" data-id="take">}\href{../../unstruwwel/html/Period.html#method-take}{\code{unstruwwel::Period$take()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Helper function to specify the beginning of a decade.


Helper function to specify the middle of a decade.


Helper function to specify the end of a decade.


Create a decade.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decade$new(value, official_def = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numerical scalar.}

\item{\code{official_def}}{If `TRUE`, the official definition that
begins with the year 1 is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object of \code{\link{R6Class}} with methods to set
common time periods and specifications for decades.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decade$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
