% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{languages}
\alias{languages}
\title{Language Information}
\format{A tibble with 4 rows and 5 variables.}
\usage{
data(languages)
}
\description{
A dataset containing the names, date orders, simplifications,
and replacements of 4 languages.
}
\keyword{datasets}
