% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-period.R
\docType{class}
\name{Period}
\alias{Period}
\title{Set a Period}
\format{An \code{\link{R6Class}} object.}
\usage{
Period
}
\value{
Object of \code{\link{R6Class}}.
}
\description{
Set a Period
}
\section{Fields}{

\describe{
\item{\code{interval}}{Stores a time interval.}
}}

\keyword{date}
\keyword{time}
