% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-year.R
\docType{class}
\name{Year}
\alias{Year}
\title{Set a Year and Get its Time Interval}
\description{
Set a Year and Get its Time Interval

Set a Year and Get its Time Interval
}
\details{
An Object of \code{R6Class} with methods to set
common time periods and specifications for years.
}
\examples{
if (interactive()) {
x <- Year$new(1520)
x$take(15, type = "june")
}

}
\section{Super class}{
\code{\link[unstruwwel:Periods]{unstruwwel::Periods}} -> \code{Year}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Year-new}{\code{Year$new()}}
\item \href{#method-Year-take}{\code{Year$take()}}
\item \href{#method-Year-clone}{\code{Year$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="unstruwwel" data-topic="Periods" data-id="set_additions"><a href='../../unstruwwel/html/Periods.html#method-Periods-set_additions'><code>unstruwwel::Periods$set_additions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Year-new"></a>}}
\if{latex}{\out{\hypertarget{method-Year-new}{}}}
\subsection{Method \code{new()}}{
Helper function to specify a time period.


Helper function to specify a season.


Helper function to specify a month.


Create a year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Year$new(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numerical scalar.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object of \code{R6Class} with methods to set
common time periods and specifications for years.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Year-take"></a>}}
\if{latex}{\out{\hypertarget{method-Year-take}{}}}
\subsection{Method \code{take()}}{
Specify a year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Year$take(x = NA, type = NA, ignore_errors = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A numerical scalar. The range of valid values
depends on \code{type}. If \code{type} is \code{"spring"},
\code{"summer"}, \code{"autumn"}, or \code{"winter"},
\code{x} is ignored.}

\item{\code{type}}{A character scalar. The following values are
supported: \code{"spring"}, \code{"summer"}, \code{"autumn"},
\code{"winter"}, and all English-language months.}

\item{\code{ignore_errors}}{If `TRUE`, error messages are ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object of \code{R6Class} with methods to set
common time periods and specifications for years.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Year-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Year-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Year$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
